NJ Colleges And Map App
This Android app, developed in Kotlin, showcases three fragments to provide comprehensive information about colleges in New Jersey. The first fragment displays all the colleges on a Google Map, utilizing the Google Maps API and markers. The app requests location permission to display where the user is located on the map as well as calculate the user's distance from each college - displaying this information on each marker in the map fragment.The second fragment features a RecyclerView list of colleges, including details such as names, towns, and addresses. Lastly, the third fragment presents a summary, including the count of colleges in North, Central, and South Jersey. Features such as a compass, zooming options, and descriptive icons are included. This app seamlessly combines mapping, listing, and summarizing functionalities to offer an immersive experience for users exploring colleges in New Jersey.